Fixes after files moved
